--- lib/PATENTS
@@ -1,22 +1,0 @@
-Additional IP Rights Grant (Patents)
-
-"This implementation" means the copyrightable works distributed by
-aerth as part of the Diamond project.
-
-aerth hereby grants to You a perpetual, worldwide, non-exclusive,
-no-charge, royalty-free, irrevocable (except as stated in this section)
-patent license to make, have made, use, offer to sell, sell, import,
-transfer and otherwise run, modify and propagate the contents of this
-implementation of Diamond, where such license applies only to those patent
-claims, both currently owned or controlled by aerth and acquired in
-the future, licensable by aerth that are necessarily infringed by this
-implementation of Diamond.  This grant does not include claims that would be
-infringed only as a consequence of further modification of this
-implementation.  If you or your agent or exclusive licensee institute or
-order or agree to the institution of patent litigation against any
-entity (including a cross-claim or counterclaim in a lawsuit) alleging
-that this implementation of Diamond or any code incorporated within this
-implementation of Diamond constitutes direct or contributory patent
-infringement, or inducement of patent infringement, then any patent
-rights granted to you under this License for this implementation of Diamond
-shall terminate as of the date such litigation is filed.
